Subscription System
Contains a base class "Subscribable" for objects to which a
subscription can be made. Subscriptions are made using the
SubscriptionPointer class. Once a subscription is made, the subscriber,
i.e. the SubscriptionPointer, has to be destroyed before the class
subscribed to (i.e. the child of Subscribable). Otherwise an exception
is thrown once the Subscribable is destroyed.
In other words the SubscriptionPointer cannot be dangling.

SubscriptionPointer represents a pointer, i.e. we can do all sort of
pointer operations with it, but we are guaranteed that it will
dereference in case it is not empty, because of the above mechanism.