Implementation BlockDiagonalMatrix class
BlockDiagonalMatrix is an implementation of a general
BlockDiagonalMatrix, which is not restricted to the type or the size or
the number of the matrix blocks along the diagonal: Any kind of lazy or
stored matrix is allowed. So we could combine different stored matrices
or lazy and stored matrices in the same BlockDiagonalMatrix object.

At the moment only rudimentary tests are done and the implementation
should not yet be trusted.