Restructured cmake buildsystem
Introduced a module called DebugReleaseBuild in order to perform the
parallel build of a debug and a release version. This module provides
some custom versions of cmake functions like add_executable,
add_library and so on in order to simplify the setup.

The module is now used in this build system and automatically installed
to the cmake module directory as well.