Basic IO functionality for formatted matrix printing
Matrix printing is now performed in the io::MatrixPrinter
The idea is to extend this and perform sensible printing
such that as much information is displayed as detailed
as possible to a user on a cout or cerr stream.
Matrix_i now uses this to implement the << operator.
The intention is that this should be used for printing
to stdout/stderr.

For printing to files the DataOut_i interface and the FileTypeWriter is
available. The only currently supported output format is Mathematica.